backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]==2.5.3

# Email
emails==0.6
aiosmtplib==3.0.1

# LINE Integration
line-bot-sdk==3.8.0

# HTTP Client
httpx==0.26.0
aiohttp>=3.9.1

# File Upload
python-magic==0.4.27
pillow==10.2.0

# Validation
pydantic-settings==2.1.0
email-validator==2.1.0

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Monitoring & Logging
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.39.2

# Rate Limiting
slowapi==0.1.9

# Environment Variables
python-dotenv==1.0.0

